plugins: add openchannel2 plugin
Adds a `openchannel2` plugin hook for channel opens that are
initiated by a peer using open_channel2.

Passes a `our_max_funding_satoshis` value in the hook to the
plugin, which can either accept or reject the channel open.

If it accepts it, plugin can specify an amount to continue with.

```
{ "result": "continue", "funding_sats": 1000000}
```

See additional details in doc/PLUGINs.md