df: just check that the funding is above what's required
we don't know what change is going to be yet, so doing an
exact match on inputs to funding here is pretty much bound
to fail.

this would be correct if we asked the funding amount to always
include fees, with the understanding that we'd take them out
but that could be weird?