df/wallet: add expiring leases to output reservations
dual funding means that we'll now have utxos that have been
shared but not spent (this means they might become spent, but
they're not longer eligible to be 'unreserved').

this patch adds the ability to track 'reserved' utxos and to record
the blockheight that they are 'reserved' at, along with a lease
time at which they'll be considered no longer reserved.