funding: track funding inputs until spent
If an input to a channel open is spent elsewhere, we channel as 'BORKED'
and wait until the 'borking' transaction has been sunk to a depth of 6;
wherein we forget the channel entirely and remove all traces of it from
the database.

Currently, we don't notify the peer of an error when the channel has
been sunk appropriately, as we expect them to also be tracking for
channel 'borking's and clean up the channel on their own behalf.

This removes/changes the way that tracking was being done; now every
channel tracks all of its own inputs and updates them using txo watches
and a tx watch for the borked transaction.

Since this is different from the first attempt at tracking outputs via
the database, we remove the test that tracked channel closure/deletions.

A few hooks have been added to provide the ability to update a channel's
status given certain actions; in the RBF open case we'll need to do more
complicated accounting of what input/tx spends permanently bork a channel
or open it.

Still to come: repopulation of watches at start/restart.