borked: re-instate info about borked transactions at start/restart
On restart, we need to re-load all of the 'borked' channel information,
so that we can accurately move the channel into the correct state as
they move forward.

Note that this assumes that the rescan done at start is >=6 from the tip.
If a rescan is done of less than 6 then borked transactions will not
be properly watched and borked channels will not get cleaned up.